webapp/requirements.txt
eventlet==0.21.0
Flask==0.11.1
Flask-Cors==2.1.2
Flask-SocketIO==2.9.0
kafka-python==1.3.3
nltk==3.2.4
python-socketio==1.7.4
tweepy==3.5.0
